# src/infrastructure/http/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arrow Maze API",
    "version": "0.2.0",
    "description": "API REST para Arrow Maze: autenticación, progreso, leaderboard y niveles."
  },
  "paths": {
...
    "/levels/{id}": {
      "get": {
        "summary": "Obtiene un nivel por id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Definición del nivel" },
          "404": { "description": "El nivel no existe" }
        }
      },
      "put": {
        "summary": "Crea o actualiza la definición de un nivel",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Nivel guardado" },
          "422": { "description": "El nivel no es resoluble" }
        }
      }
    }
  }
}